add resource files for testing
